[feature] CI on nvhpc, the NVIDIA HPC compilers (#237)
Update excit to remove issues with hardcoded -Werror, test the code
against NVHPC SDK 24.7, and fix a couple of bugs in the process.

Note: src/features.c depends on CUDA when CUDA is active, so might as
well include the CUDA flags in the entire library.

Note: The hwloc tests are written to test hwloc or the platform, instead of
testing that aml is doing the right thing. We will need to change them
drastically.

+++ benchmarks/blas/l1_kernel.c
@@ -56,15 +56,18 @@ double ddot(size_t n, double *a, double *b, double *c, double scalar)
 	(void)*c;
 	(void)scalar;
 	size_t i;
-	long double dot = 0.0;
+	/* should be a long double for overflow checks, but some compilers (nvc)
+	 * don't support reduce on long double in 2024.
+	 */
+	double dot = 0.0;
 
 #pragma omp parallel for reduction(+ : dot)
 	for (i = 0; i < n; i++) {
-		long double temp;
+		double temp;
 		temp = a[i] * b[i];
 		dot += temp;
 	}
-	return (double)dot;
+	return dot;
 }
 
 double dnrm2(size_t n, double *a, double *b, double *c, double scalar)
